library/requirements.txt
numpy
biopython
leuci-xyz
pandas
matplotlib
seaborn
scipy
ruff
pytest
twine
build
urllib3==1.26.6